Started styling .info section. Changed the size of the main content. Set position to nav bar to fixed.
Changed the background of the info element to a the same color as the hover color in the nav bar. Added padding on each div within the info element and centered them. Added icons to each of the three divs within .info. Changed the neight of the main content to take up more of the vh of the screen.